requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Free financial data libraries
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Optional: FRED API (free registration required)
fredapi>=0.5.0

# Free crypto data
requests>=2.31.0
beautifulsoup4>=4.12.0

# Statistical analysis
statsmodels>=0.14.0
arch>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Documentation
sphinx>=7.0.0